backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
razorpay==1.4.1
pydantic-settings==2.4.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
httpx==0.27.0


